chore: extract McpFacade from CamelMonitor to reduce God class
Move ~48 MCP-facing accessor methods (~650 lines) from CamelMonitor
into a new McpFacade class. TuiMcpServer now depends only on
McpFacade, decoupling it from the full monitor. This is step 1 of
the CamelMonitor decomposition plan.

Key changes:
- New McpFacade.java with all MCP accessor, navigation, data, and
  control methods plus TAB_NAMES/MORE_TAB_NAMES constants and the
  PendingKey record
- MonitorBridge callback interface to decouple McpFacade from
  CamelMonitor internals
- AtomicReference<TapeRecorder> for thread-safe sharing between
  monitor event loop and MCP facade
- TuiMcpServer updated to take McpFacade instead of CamelMonitor
- CamelMonitorParseKeyTest updated for moved parseKey() method
- All 260 tests pass, zero behavior change
