+++ dsl/camel-jbang/camel-jbang-plugin-tui/src/main/java/org/apache/camel/dsl/jbang/core/commands/tui/CamelMonitor.java
@@ -29,7 +29,6 @@ import java.util.Collections;
 import java.util.Comparator;
 import java.util.Iterator;
 import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Queue;
@@ -58,7 +57,6 @@ import dev.tamboui.tui.TuiRunner;
 import dev.tamboui.tui.event.Event;
 import dev.tamboui.tui.event.KeyCode;
 import dev.tamboui.tui.event.KeyEvent;
-import dev.tamboui.tui.event.KeyModifiers;
 import dev.tamboui.tui.event.MouseEvent;
 import dev.tamboui.tui.event.PasteEvent;
 import dev.tamboui.tui.event.TickEvent;
@@ -78,7 +76,6 @@ import org.apache.camel.dsl.jbang.core.commands.CamelCommand;
 import org.apache.camel.dsl.jbang.core.commands.CamelJBangMain;
 import org.apache.camel.dsl.jbang.core.common.CommandLineHelper;
 import org.apache.camel.dsl.jbang.core.common.PathUtils;
-import org.apache.camel.dsl.jbang.core.common.RuntimeHelper;
 import org.apache.camel.dsl.jbang.core.common.VersionHelper;
 import org.apache.camel.util.json.JsonArray;
 import org.apache.camel.util.json.JsonObject;
@@ -175,11 +172,12 @@ public class CamelMonitor extends CamelCommand {
     private volatile long screenshotMessageTime;
     private volatile boolean pendingScreenshot;
     private boolean recording;
-    private TapeRecorder tapeRecorder;
+    private final AtomicReference<TapeRecorder> tapeRecorderRef = new AtomicReference<>();
     private boolean mcpInjectedKey;
     private TuiEventLog eventLog;
     private TuiMcpServer mcpServer;
-    private final Queue<PendingKey> pendingKeys = new ConcurrentLinkedQueue<>();
+    private McpFacade mcpFacade;
+    private final Queue<McpFacade.PendingKey> pendingKeys = new ConcurrentLinkedQueue<>();
     private final List<KeyRecord> recentKeys = new ArrayList<>();
     private final CaptionOverlay captionOverlay = new CaptionOverlay();
     private final DrawOverlay drawOverlay = new DrawOverlay();
@@ -202,7 +200,10 @@ public class CamelMonitor extends CamelCommand {
             () -> recording = !recording,
             () -> recording,
             this::toggleTapeRecording,
-            () -> tapeRecorder != null && tapeRecorder.isActive(),
+            () -> {
+                TapeRecorder r = tapeRecorderRef.get();
+                return r != null && r.isActive();
+            },
             this::enableBurstMode, stoppingPids);
 
     private final AtomicBoolean refreshInProgress = new AtomicBoolean(false);
@@ -316,9 +317,81 @@ public class CamelMonitor extends CamelCommand {
         overviewTab.selectCurrentIntegration();
 
         eventLog = new TuiEventLog(500);
+        mcpFacade = new McpFacade(
+                ctx, data, tabsState, eventLog,
+                captionOverlay, drawOverlay, helpOverlay,
+                actionsPopup, filesBrowser,
+                logTab, diagramTab, historyTab,
+                tapeRecorderRef, pendingKeys,
+                new McpFacade.MonitorBridge() {
+                    @Override
+                    public MonitorTab activeTab() {
+                        return CamelMonitor.this.activeTab();
+                    }
+
+                    @Override
+                    public Buffer lastBuffer() {
+                        return lastBuffer;
+                    }
+
+                    @Override
+                    public long renderGeneration() {
+                        return renderGeneration;
+                    }
+
+                    @Override
+                    public boolean isKeystrokesVisible() {
+                        return recording;
+                    }
+
+                    @Override
+                    public void handleTabKey(int tabIndex) {
+                        CamelMonitor.this.handleTabKey(tabIndex);
+                    }
+
+                    @Override
+                    public void selectMoreTab(int moreIndex) {
+                        CamelMonitor.this.selectMoreTab(moreIndex);
+                    }
+
+                    @Override
+                    public boolean isSwitchPopupVisible() {
+                        return showSwitchPopup;
+                    }
+
+                    @Override
+                    public boolean isMorePopupVisible() {
+                        return showMorePopup;
+                    }
+
+                    @Override
+                    public void renderOverviewFooter(List<Span> spans) {
+                        CamelMonitor.this.renderOverviewFooter(spans);
+                    }
+
+                    @Override
+                    public void insertFKeyHints(List<Span> spans) {
+                        CamelMonitor.this.insertFKeyHints(spans);
+                    }
+
+                    @Override
+                    public void sendRouteCommand(String pid, String routeId, String command) {
+                        CamelMonitor.this.sendRouteCommand(pid, routeId, command);
+                    }
+
+                    @Override
+                    public void restartProcess() {
+                        restartSelectedProcess();
+                    }
+
+                    @Override
+                    public void stopProcess(boolean forceKill) {
+                        stopSelectedProcess(forceKill);
+                    }
+                });
         Path mcpJsonFile = null;
         if (mcp) {
-            mcpServer = new TuiMcpServer(mcpPort, this);
+            mcpServer = new TuiMcpServer(mcpPort, mcpFacade);
             try {
                 mcpServer.start();
                 actionsPopup.setMcpEnabled(true, mcpPort, mcpServer::getConnectedClient, mcpServer::getActivityLog);
@@ -376,10 +449,11 @@ public class CamelMonitor extends CamelCommand {
                 toggleTapeRecording();
                 return true;
             }
-            if (tapeRecorder != null && tapeRecorder.isActive() && !mcpInjectedKey) {
+            TapeRecorder tr = tapeRecorderRef.get();
+            if (tr != null && tr.isActive() && !mcpInjectedKey) {
                 String label = keyLabel(ke);
                 if (label != null) {
-                    tapeRecorder.recordKey(label);
+                    tr.recordKey(label);
                 }
             }
             if (captionOverlay.isVisible()) {
@@ -458,30 +532,7 @@ public class CamelMonitor extends CamelCommand {
                 showMorePopup = false;
                 Integer sel = shortcutSel >= 0 ? shortcutSel : morePopupState.selected();
                 if (sel != null) {
-                    lastMoreSelection = sel;
-                    activeMoreTab = switch (sel) {
-                        case 0 -> beansTab;
-                        case 1 -> browseTab;
-                        case 2 -> circuitBreakerTab;
-                        case 3 -> classpathTab;
-                        case 4 -> configurationTab;
-                        case 5 -> consumersTab;
-                        case 6 -> dataSourceTab;
-                        case 7 -> inflightTab;
-                        case 8 -> memoryTab;
-                        case 9 -> metricsTab;
-                        case 10 -> sqlQueryTab;
-                        case 11 -> spansTab;
-                        case 12 -> processTab;
-                        case 13 -> startupTab;
-                        case 14 -> threadsTab;
-                        default -> null;
-                    };
-                    if (activeMoreTab != null) {
-                        overviewTab.selectCurrentIntegration();
-                        tabsState.select(TAB_MORE);
-                        activeMoreTab.onTabSelected();
-                    }
+                    selectMoreTab(sel);
                 }
                 return true;
             }
@@ -799,7 +850,7 @@ public class CamelMonitor extends CamelCommand {
     private boolean handleTickEvent(TuiRunner runner) {
         long now = System.currentTimeMillis();
         boolean keyProcessed = false;
-        PendingKey pk;
+        McpFacade.PendingKey pk;
         while ((pk = pendingKeys.peek()) != null && now >= pk.fireAt()) {
             pendingKeys.poll();
             mcpInjectedKey = true;
@@ -932,6 +983,34 @@ public class CamelMonitor extends CamelCommand {
         showMorePopup = false;
         tabsState.select(tab);
         return true;
+    }
+
+    void selectMoreTab(int index) {
+        morePopupState.select(index);
+        lastMoreSelection = index;
+        activeMoreTab = switch (index) {
+            case 0 -> beansTab;
+            case 1 -> browseTab;
+            case 2 -> circuitBreakerTab;
+            case 3 -> classpathTab;
+            case 4 -> configurationTab;
+            case 5 -> consumersTab;
+            case 6 -> dataSourceTab;
+            case 7 -> inflightTab;
+            case 8 -> memoryTab;
+            case 9 -> metricsTab;
+            case 10 -> sqlQueryTab;
+            case 11 -> spansTab;
+            case 12 -> processTab;
+            case 13 -> startupTab;
+            case 14 -> threadsTab;
+            default -> null;
+        };
+        if (activeMoreTab != null) {
+            overviewTab.selectCurrentIntegration();
+            tabsState.select(TAB_MORE);
+            activeMoreTab.onTabSelected();
+        }
     }
 
     private List<Long> selectedPidAsList() {
@@ -2489,46 +2568,11 @@ public class CamelMonitor extends CamelCommand {
         }
     }
 
-    // ---- MCP accessor methods ----
-
-    private static final String[] TAB_NAMES = {
-            "Overview", "Log", "Diagram", "Routes", "Endpoints",
-            "HTTP", "Health", "Inspect", "Errors", "More"
-    };
-
-    Buffer getLastBuffer() {
-        return lastBuffer;
-    }
-
-    long getRenderGeneration() {
-        return renderGeneration;
-    }
-
-    boolean isKeystrokesVisible() {
-        return recording;
-    }
-
-    TapeRecorder getTapeRecorder() {
-        return tapeRecorder;
-    }
-
-    boolean isTapeRecording() {
-        return tapeRecorder != null && tapeRecorder.isActive();
-    }
-
-    void startTapeRecording(String title) {
-        tapeRecorder = new TapeRecorder();
-        tapeRecorder.start(title);
-    }
-
-    void clearTapeRecorder() {
-        tapeRecorder = null;
-    }
-
     private void toggleTapeRecording() {
-        if (tapeRecorder != null && tapeRecorder.isActive()) {
-            String tape = tapeRecorder.stop();
-            tapeRecorder = null;
+        TapeRecorder rec = tapeRecorderRef.get();
+        if (rec != null && rec.isActive()) {
+            String tape = rec.stop();
+            tapeRecorderRef.set(null);
             String timestamp = java.time.LocalDateTime.now()
                     .format(java.time.format.DateTimeFormatter.ofPattern("yyyyMMdd-HHmmss"));
             String filename = "camel-tui-tape-" + timestamp + ".tape";
@@ -2539,662 +2583,11 @@ public class CamelMonitor extends CamelCommand {
                 captionOverlay.showCaption("Failed to save tape: " + e.getMessage(), 5);
             }
         } else {
-            tapeRecorder = new TapeRecorder();
-            tapeRecorder.start(null);
+            rec = new TapeRecorder();
+            rec.start(null);
+            tapeRecorderRef.set(rec);
             captionOverlay.showCaption("Tape recording started", 3);
         }
-    }
-
-    TuiEventLog getEventLog() {
-        return eventLog;
-    }
-
-    int getActiveTabIndex() {
-        return tabsState.selected();
-    }
-
-    String getActiveTabName() {
-        int idx = tabsState.selected();
-        return idx >= 0 && idx < TAB_NAMES.length ? TAB_NAMES[idx] : "Unknown";
-    }
-
-    String getSelectedPid() {
-        return ctx != null ? ctx.selectedPid : null;
-    }
-
-    String getSelectedIntegrationName() {
-        if (ctx == null) {
-            return null;
-        }
-        IntegrationInfo info = ctx.findSelectedIntegration();
-        return info != null ? info.name : null;
-    }
-
-    int getIntegrationCount() {
-        List<IntegrationInfo> list = data.get();
-        return (int) list.stream().filter(i -> !i.vanishing).count();
-    }
-
-    boolean isCaptionVisible() {
-        return captionOverlay.isCaptionVisible();
-    }
-
-    void showCaption(String text) {
-        captionOverlay.showCaption(text);
-    }
-
-    void showCaption(String text, int durationSeconds) {
-        captionOverlay.showCaption(text, durationSeconds);
-    }
-
-    boolean isDrawVisible() {
-        return drawOverlay.isVisible();
-    }
-
-    void setDrawing(List<DrawOverlay.DrawCell> cells, int durationSeconds) {
-        drawOverlay.setDrawing(cells, durationSeconds);
-    }
-
-    void appendDrawing(List<DrawOverlay.DrawCell> cells) {
-        drawOverlay.appendDrawing(cells);
-    }
-
-    void clearDrawing() {
-        drawOverlay.clear();
-    }
-
-    private static final String[] MORE_TAB_NAMES = {
-            "Beans", "Browse", "Circuit Breaker", "Classpath", "Configuration",
-            "Consumers", "DataSource", "Inflight", "Memory", "Metrics", "SQL Query", "Spans", "Process", "Startup",
-            "Threads"
-    };
-
-    String navigateToTab(String tabName) {
-        for (int i = 0; i < TAB_NAMES.length; i++) {
-            if (TAB_NAMES[i].equalsIgnoreCase(tabName)) {
-                handleTabKey(i);
-                return TAB_NAMES[i];
-            }
-        }
-        // Check More submenu tabs
-        for (int i = 0; i < MORE_TAB_NAMES.length; i++) {
-            if (MORE_TAB_NAMES[i].equalsIgnoreCase(tabName)) {
-                morePopupState.select(i);
-                lastMoreSelection = i;
-                activeMoreTab = switch (i) {
-                    case 0 -> beansTab;
-                    case 1 -> browseTab;
-                    case 2 -> circuitBreakerTab;
-                    case 3 -> classpathTab;
-                    case 4 -> configurationTab;
-                    case 5 -> consumersTab;
-                    case 6 -> dataSourceTab;
-                    case 7 -> inflightTab;
-                    case 8 -> memoryTab;
-                    case 9 -> metricsTab;
-                    case 10 -> sqlQueryTab;
-                    case 11 -> spansTab;
-                    case 12 -> processTab;
-                    case 13 -> startupTab;
-                    case 14 -> threadsTab;
-                    default -> null;
-                };
-                if (activeMoreTab != null) {
-                    overviewTab.selectCurrentIntegration();
-                    tabsState.select(TAB_MORE);
-                    activeMoreTab.onTabSelected();
-                }
-                return MORE_TAB_NAMES[i];
-            }
-        }
-        return null;
-    }
-
-    String selectIntegration(String nameOrPid) {
-        List<IntegrationInfo> infos = data.get();
-        for (IntegrationInfo info : infos) {
-            if (info.vanishing) {
-                continue;
-            }
-            if (nameOrPid.equals(info.pid)
-                    || (info.name != null && info.name.equalsIgnoreCase(nameOrPid))) {
-                ctx.selectedPid = info.pid;
-                return info.name != null ? info.name : info.pid;
-            }
-        }
-        return null;
-    }
-
-    List<String> getTabNames() {
-        List<String> names = new ArrayList<>();
-        names.addAll(List.of(TAB_NAMES));
-        names.addAll(List.of(MORE_TAB_NAMES));
-        return names;
-    }
-
-    List<String> getActionLabels() {
-        return actionsPopup.getActionLabels();
-    }
-
-    SelectionContext getSelectionContext() {
-        SelectionContext popup = actionsPopup.getSelectionContext();
-        if (popup != null) {
-            return popup;
-        }
-        MonitorTab tab = activeTab();
-        return tab != null ? tab.getSelectionContext() : null;
-    }
-
-    List<String> getIntegrationNames() {
-        return data.get().stream()
-                .filter(i -> !i.vanishing)
-                .map(i -> i.name != null ? i.name : i.pid)
-                .toList();
-    }
-
-    JsonObject getReadme(String name) {
-        List<IntegrationInfo> integrations = data.get();
-        IntegrationInfo target = null;
-        if (name != null && !name.isEmpty()) {
-            for (IntegrationInfo info : integrations) {
-                if (!info.vanishing && name.equals(info.name)) {
-                    target = info;
-                    break;
-                }
-            }
-        } else {
-            target = ctx != null ? ctx.findSelectedIntegration() : null;
-        }
-        if (target == null) {
-            return null;
-        }
-        if (target.readmeFiles == null || target.readmeFiles.isEmpty()) {
-            return null;
-        }
-        try {
-            Path outputFile = ctx.getOutputFile(target.pid);
-            Files.deleteIfExists(outputFile);
-            JsonObject action = new JsonObject();
-            action.put("action", "readme");
-            PathUtils.writeTextSafely(action.toJson(), ctx.getActionFile(target.pid));
-            return MonitorContext.pollJsonResponse(outputFile, 5000);
-        } catch (Exception e) {
-            return null;
-        }
-    }
-
-    JsonObject getFiles(String name, String file) {
-        List<IntegrationInfo> integrations = data.get();
-        IntegrationInfo target = null;
-        if (name != null && !name.isEmpty()) {
-            for (IntegrationInfo info : integrations) {
-                if (!info.vanishing && name.equals(info.name)) {
-                    target = info;
-                    break;
-                }
-            }
-        } else {
-            target = ctx != null ? ctx.findSelectedIntegration() : null;
-        }
-        if (target == null) {
-            return null;
-        }
-        Path dir = FilesBrowser.resolveSourceDirectory(target);
-        if (dir == null || !Files.isDirectory(dir)) {
-            return null;
-        }
-        if (file != null && !file.isEmpty()) {
-            Path filePath = dir.resolve(file);
-            if (!Files.isRegularFile(filePath)) {
-                return null;
-            }
-            try {
-                String content = Files.readString(filePath, StandardCharsets.UTF_8);
-                JsonObject result = new JsonObject();
-                result.put("file", file);
-                result.put("directory", dir.toString());
-                result.put("size", FilesBrowser.formatFileSize(Files.size(filePath)));
-                result.put("type", FilesBrowser.fileType(filePath));
-                result.put("content", content);
-                return result;
-            } catch (IOException e) {
-                return null;
-            }
-        }
-        JsonArray files = new JsonArray();
-        try (var stream = Files.list(dir)) {
-            stream.filter(Files::isRegularFile)
-                    .sorted((a, b) -> a.getFileName().toString().compareToIgnoreCase(b.getFileName().toString()))
-                    .limit(99)
-                    .forEach(p -> {
-                        JsonObject entry = new JsonObject();
-                        entry.put("name", p.getFileName().toString());
-                        try {
-                            entry.put("size", FilesBrowser.formatFileSize(Files.size(p)));
-                        } catch (IOException e) {
-                            entry.put("size", "0 B");
-                        }
-                        entry.put("type", FilesBrowser.fileType(p));
-                        files.add(entry);
-                    });
-        } catch (IOException e) {
-            return null;
-        }
-        if (files.isEmpty()) {
-            return null;
-        }
-        JsonObject result = new JsonObject();
-        result.put("directory", dir.toString());
-        result.put("files", files);
-        result.put("totalFiles", files.size());
-        return result;
-    }
-
-    int injectKeys(List<String> keys, int delayMs) {
-        long fireAt = System.currentTimeMillis();
-        int count = 0;
-        for (String key : keys) {
-            KeyEvent ke = parseKey(key);
-            if (ke != null) {
-                pendingKeys.add(new PendingKey(ke, fireAt));
-                fireAt += delayMs;
-                count++;
-            }
-        }
-        return count;
-    }
-
-    static KeyEvent parseKey(String key) {
-        if (key == null || key.isEmpty()) {
-            return null;
-        }
-
-        boolean ctrl = false;
-        boolean shift = false;
-        String remainder = key;
-        while (remainder.contains("+")) {
-            int plus = remainder.indexOf('+');
-            String mod = remainder.substring(0, plus).trim();
-            remainder = remainder.substring(plus + 1).trim();
-            if (mod.equalsIgnoreCase("Ctrl")) {
-                ctrl = true;
-            } else if (mod.equalsIgnoreCase("Shift")) {
-                shift = true;
-            }
-        }
-
-        KeyModifiers mods = KeyModifiers.of(ctrl, false, shift);
-
-        KeyCode code = switch (remainder.toLowerCase(Locale.ROOT)) {
-            case "enter", "return" -> KeyCode.ENTER;
-            case "esc", "escape" -> KeyCode.ESCAPE;
-            case "tab" -> KeyCode.TAB;
-            case "backspace" -> KeyCode.BACKSPACE;
-            case "delete", "del" -> KeyCode.DELETE;
-            case "up" -> KeyCode.UP;
-            case "down" -> KeyCode.DOWN;
-            case "left" -> KeyCode.LEFT;
-            case "right" -> KeyCode.RIGHT;
-            case "home" -> KeyCode.HOME;
-            case "end" -> KeyCode.END;
-            case "pageup", "pgup" -> KeyCode.PAGE_UP;
-            case "pagedown", "pgdn" -> KeyCode.PAGE_DOWN;
-            case "f1" -> KeyCode.F1;
-            case "f2" -> KeyCode.F2;
-            case "f3" -> KeyCode.F3;
-            case "f4" -> KeyCode.F4;
-            case "f5" -> KeyCode.F5;
-            case "f6" -> KeyCode.F6;
-            case "f7" -> KeyCode.F7;
-            case "f8" -> KeyCode.F8;
-            case "f9" -> KeyCode.F9;
-            case "f10" -> KeyCode.F10;
-            case "f11" -> KeyCode.F11;
-            case "f12" -> KeyCode.F12;
-            case "space" -> null;
-            default -> null;
-        };
-
-        if (code != null) {
-            return KeyEvent.ofKey(code, mods);
-        }
-        if ("space".equalsIgnoreCase(remainder)) {
-            return KeyEvent.ofChar(' ', mods);
-        }
-        if (remainder.length() == 1) {
-            return KeyEvent.ofChar(remainder.charAt(0), mods);
-        }
-        return null;
-    }
-
-    JsonObject getTableData(String tabName) {
-        if (tabName != null && !tabName.isBlank()) {
-            String prev = getActiveTabName();
-            String switched = navigateToTab(tabName);
-            if (switched == null) {
-                return null;
-            }
-        }
-        MonitorTab tab = activeTab();
-        return tab != null ? tab.getTableDataAsJson() : null;
-    }
-
-    boolean executeAction(String actionName) {
-        return actionsPopup.executeActionByName(actionName);
-    }
-
-    JsonObject getLogData(int limit, String filter, String level) {
-        return logTab.getLogDataAsJson(limit, filter, level);
-    }
-
-    JsonObject getDiagramData() {
-        MonitorTab tab = activeTab();
-        if (tab instanceof DiagramTab dt) {
-            return dt.getTableDataAsJson();
-        }
-        return diagramTab.getTableDataAsJson();
-    }
-
-    void selectTraceExchange(String exchangeId) {
-        historyTab.selectTraceExchange(exchangeId);
-    }
-
-    JsonObject getTopologyData() {
-        return diagramTab.getTopologyDataAsJson();
-    }
-
-    @SuppressWarnings("unchecked")
-    JsonObject getSpanData(String traceId, int limit) {
-        String pid = ctx.selectedPid;
-        if (pid == null) {
-            JsonObject err = new JsonObject();
-            err.put("error", "No integration selected");
-            return err;
-        }
-        try {
-            Path outputFile = ctx.getOutputFile(pid);
-            PathUtils.deleteFile(outputFile);
-
-            JsonObject action = new JsonObject();
-            action.put("action", "span");
-            action.put("dump", "true");
-            action.put("limit", String.valueOf(limit));
-            Path actionFile = ctx.getActionFile(pid);
-            PathUtils.writeTextSafely(action.toJson(), actionFile);
-
-            JsonObject response = MonitorContext.pollJsonResponse(outputFile, 3000);
-            if (response != null) {
-                PathUtils.deleteFile(outputFile);
-                Boolean enabled = response.getBoolean("enabled");
-                if (enabled == null || !enabled) {
-                    JsonObject err = new JsonObject();
-                    err.put("error", "OpenTelemetry not enabled (requires --observe flag)");
-                    return err;
-                }
-                if (traceId != null && !traceId.isBlank()) {
-                    JsonArray all = response.getCollection("spans");
-                    if (all != null) {
-                        JsonArray filtered = new JsonArray();
-                        for (int i = 0; i < all.size(); i++) {
-                            JsonObject span = (JsonObject) all.get(i);
-                            String tid = span.getString("traceId");
-                            if (tid != null && tid.contains(traceId)) {
-                                filtered.add(span);
-                            }
-                        }
-                        response.put("spans", filtered);
-                    }
-                }
-                return response;
-            }
-            JsonObject err = new JsonObject();
-            err.put("error", "Timeout waiting for span data");
-            return err;
-        } catch (Exception e) {
-            JsonObject err = new JsonObject();
-            err.put("error", e.getMessage());
-            return err;
-        }
-    }
-
-    String navigateDiagramToRoute(String routeId) {
-        navigateToTab("Diagram");
-        if (diagramTab.selectRoute(routeId)) {
-            return routeId;
-        }
-        return null;
-    }
-
-    String navigateDiagramToNode(String routeId, String nodeId) {
-        navigateToTab("Diagram");
-        if (diagramTab.selectNode(routeId, nodeId)) {
-            return nodeId;
-        }
-        return null;
-    }
-
-    JsonObject getDiagramState() {
-        return diagramTab.getDiagramStateAsJson();
-    }
-
-    JsonArray locateText(String search) {
-        Buffer buf = lastBuffer;
-        if (buf == null || search == null || search.isEmpty()) {
-            return new JsonArray();
-        }
-        String screen = ExportRequest.export(buf).text().toString();
-        String[] lines = screen.split("\n", -1);
-        int searchWidth = 0;
-        for (int i = 0; i < search.length();) {
-            int cp = search.codePointAt(i);
-            searchWidth += Math.max(1, CharWidth.of(cp));
-            i += Character.charCount(cp);
-        }
-        JsonArray matches = new JsonArray();
-        for (int y = 0; y < lines.length; y++) {
-            String line = lines[y];
-            int idx = line.indexOf(search);
-            while (idx >= 0) {
-                int visualCol = 0;
-                for (int i = 0; i < idx;) {
-                    int cp = line.codePointAt(i);
-                    visualCol += Math.max(1, CharWidth.of(cp));
-                    i += Character.charCount(cp);
-                }
-                JsonObject match = new JsonObject();
-                match.put("x", visualCol);
-                match.put("y", y);
-                match.put("width", searchWidth);
-                match.put("height", 1);
-                match.put("text", search);
-                matches.add(match);
-                idx = line.indexOf(search, idx + 1);
-            }
-            if (matches.size() >= 20) {
-                break;
-            }
-        }
-        return matches;
-    }
-
-    JsonObject locateNodes(List<String> nodeIds) {
-        return diagramTab.locateNodes(nodeIds);
-    }
-
-    JsonArray getFooterActionsAsJson() {
-        List<Span> spans = new ArrayList<>();
-        if (helpOverlay.isVisible()) {
-            helpOverlay.renderFooter(spans);
-        } else if (captionOverlay.isCaptionVisible()) {
-            captionOverlay.renderFooter(spans);
-        } else if (filesBrowser.isVisible()) {
-            filesBrowser.renderFooter(spans);
-        } else if (showSwitchPopup || showMorePopup) {
-            if (showSwitchPopup) {
-                hint(spans, "Up/Down", "select");
-                hint(spans, "Enter", "switch");
-                hint(spans, "Esc", "close");
-            } else {
-                hint(spans, "Up/Down", "select");
-                hint(spans, "Enter", "open");
-                hint(spans, "Esc", "close");
-            }
-        } else {
-            MonitorTab tab = activeTab();
-            if (tabsState.selected() == TAB_OVERVIEW) {
-                renderOverviewFooter(spans);
-            } else if (tab != null) {
-                tab.renderFooter(spans);
-                insertFKeyHints(spans);
-            }
-        }
-        JsonArray actions = new JsonArray();
-        for (int i = 0; i + 1 < spans.size(); i += 2) {
-            String key = spans.get(i).content().trim();
-            String rawLabel = spans.get(i + 1).content().trim();
-            // compact "show BHPV" pattern: key="show", then space, then 4 single-letter spans, then trailing space
-            if ("show".equals(key) && i + 6 < spans.size()) {
-                for (int j = 0; j < 4; j++) {
-                    Span letter = spans.get(i + 2 + j);
-                    String ch = letter.content();
-                    boolean on = ch.equals(ch.toUpperCase());
-                    JsonObject toggle = new JsonObject();
-                    toggle.put("key", ch.toLowerCase());
-                    String label = switch (ch.toLowerCase()) {
-                        case "b" -> "body";
-                        case "h" -> "headers";
-                        case "p" -> "properties";
-                        case "v" -> "variables";
-                        default -> ch;
-                    };
-                    toggle.put("label", label);
-                    toggle.put("state", on ? "on" : "off");
-                    actions.add(toggle);
-                }
-                i += 5; // skip the 7-span group (loop adds 2, we consumed 5 more)
-                continue;
-            }
-            JsonObject action = new JsonObject();
-            action.put("key", key);
-            int bracket = rawLabel.indexOf('[');
-            if (bracket > 0 && rawLabel.endsWith("]")) {
-                action.put("label", rawLabel.substring(0, bracket).trim());
-                action.put("state", rawLabel.substring(bracket + 1, rawLabel.length() - 1));
-            } else {
-                action.put("label", rawLabel);
-            }
-            actions.add(action);
-        }
-        return actions;
-    }
-
-    void setLogLevel(String level) {
-        logTab.setLogLevel(level);
-    }
-
-    boolean setTabFilter(String tabName, String filter) {
-        if (tabName != null) {
-            navigateToTab(tabName);
-        }
-        MonitorTab tab = activeTab();
-        return tab != null && tab.setFilter(filter);
-    }
-
-    boolean setTabInputValue(String tabName, String field, String value) {
-        if (tabName != null) {
-            navigateToTab(tabName);
-        }
-        MonitorTab tab = activeTab();
-        return tab != null && tab.setInputValue(field, value);
-    }
-
-    String toggleTraceDisplay(String section, Boolean enabled) {
-        return historyTab.toggleDisplaySection(section, enabled);
-    }
-
-    JsonObject sendMessage(String endpoint, String body, String headers) {
-        if (ctx.selectedPid == null) {
-            return null;
-        }
-        long pid;
-        try {
-            pid = Long.parseLong(ctx.selectedPid);
-        } catch (NumberFormatException e) {
-            return null;
-        }
-        return RuntimeHelper.sendMessage(pid, endpoint, body, headers);
-    }
-
-    JsonObject executeSql(String sql, String datasource, int maxRows, int queryTimeout) {
-        if (ctx.selectedPid == null) {
-            return null;
-        }
-        long pid;
-        try {
-            pid = Long.parseLong(ctx.selectedPid);
-        } catch (NumberFormatException e) {
-            return null;
-        }
-        return RuntimeHelper.executeSqlQuery(pid, sql, datasource, maxRows, queryTimeout);
-    }
-
-    JsonObject updateRow(String table, String datasource, String pkValuesJson, String colValuesJson) {
-        if (ctx.selectedPid == null) {
-            return null;
-        }
-        long pid;
-        try {
-            pid = Long.parseLong(ctx.selectedPid);
-        } catch (NumberFormatException e) {
-            return null;
-        }
-        return RuntimeHelper.executeRowUpdate(pid, table, datasource, pkValuesJson, colValuesJson);
-    }
-
-    String controlIntegration(String action) {
-        if (action == null || action.isBlank()) {
-            return "Error: action is required";
-        }
-        if (ctx.selectedPid == null) {
-            return "Error: no integration selected";
-        }
-        String name = selectedName();
-        return switch (action) {
-            case "stop-routes", "pause" -> {
-                if (isInfraSelected()) {
-                    yield "Error: cannot stop routes on infra service";
-                }
-                sendRouteCommand(ctx.selectedPid, "*", "stop");
-                yield "Routes stopped for " + name;
-            }
-            case "start-routes", "resume" -> {
-                if (isInfraSelected()) {
-                    yield "Error: cannot start routes on infra service";
-                }
-                sendRouteCommand(ctx.selectedPid, "*", "start");
-                yield "Routes started for " + name;
-            }
-            case "restart" -> {
-                if (isInfraSelected()) {
-                    yield "Error: cannot restart infra service";
-                }
-                restartSelectedProcess();
-                yield "Restarting " + name;
-            }
-            case "stop" -> {
-                stopSelectedProcess(false);
-                yield "Stopping " + name;
-            }
-            case "kill" -> {
-                stopSelectedProcess(true);
-                yield "Killed " + name;
-            }
-            default -> "Unknown action: " + action
-                       + ". Available: stop-routes, start-routes, restart, stop, kill";
-        };
-    }
-
-    private record PendingKey(KeyEvent event, long fireAt) {
     }
 
 }
